added BridgeTransaction to own @types file
