+++ bridge-ui/src/components/bridge/transaction-history/list-transaction/index.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import styles from "./list-transaction.module.scss";
 import Transaction from "./item";
 import TransactionDetails from "@/components/bridge/transaction-history/modal/transaction-details";
-import { BridgeTransaction } from "@/utils";
+import { BridgeTransaction } from "@/types";
 
 type Props = {
   transactions: BridgeTransaction[];
