+++ bridge-ui/src/utils/index.ts
@@ -2,7 +2,7 @@ export { eventETH, eventERC20, eventERC20V2, eventUSDC } from "./events";
 export { generateChain, generateChains, getChainLogoPath, getChainNetworkLayer } from "./chains";
 export { estimateEthGasFee, estimateERC20GasFee } from "./fees";
 export { formatAddress, formatBalance, formatHex, formatTimestamp, safeGetAddress } from "./format";
-export { fetchTransactionsHistory, type BridgeTransaction } from "./history";
+export { fetchTransactionsHistory } from "./history";
 export { computeMessageHash, computeMessageStorageSlot } from "./message";
 export { isEth } from "./tokens";
 export { isEmptyObject } from "./utils";
